rename scroll watcher to main-with-ref
