+++ apps/web/app/(app)/talent/layout.tsx
@@ -1,5 +1,5 @@
 import { Nav } from './nav'
-import { MainWithRefContext } from './scroll-watcher'
+import { MainWithRefContext } from './main-with-ref'
 
 interface UserScreensLayoutProps {
   title: React.ReactNode
